fix: test can start openwhyd server or re-use running instance

+++ test/approval.tests.js
@@ -7,9 +7,12 @@ const { promisify } = require('util');
 const {
   loadEnvVars,
   startOpenwhydServer,
+  refreshOpenwhydCache,
   readMongoDocuments,
   insertTestData,
 } = require('./db-helpers');
+
+const START_WITH_ENV_FILE = process.env.START_WITH_ENV_FILE;
 
 const MONGODB_URL =
   process.env.MONGODB_URL || 'mongodb://localhost:27117/openwhyd_test';
@@ -26,28 +29,35 @@ async function getCleanedPageBody(body) {
 }
 
 test.before(async (t) => {
-  // insert test data
   const testDataCollections = {
     user: await readMongoDocuments(__dirname + '/approval.users.json'),
     post: await readMongoDocuments(__dirname + '/approval.posts.json'),
     follow: await readMongoDocuments(__dirname + '/approval.follows.json'),
   };
   await insertTestData(MONGODB_URL, testDataCollections);
-  // start server on test environment
-  const env = {
-    ...(await loadEnvVars('./env-vars-testing.conf')),
-    MONGODB_PORT: '27117', // port exposed by docker container
-    TZ: 'UTC',
-  };
-  process.env.WHYD_GENUINE_SIGNUP_SECRET = env.WHYD_GENUINE_SIGNUP_SECRET; // required by ./api-client.js
-  t.context.serverProcess = await startOpenwhydServer(env);
+
+  if (START_WITH_ENV_FILE) {
+    const env = {
+      ...(await loadEnvVars(START_WITH_ENV_FILE)),
+      MONGODB_PORT: '27117', // port exposed by docker container
+      TZ: 'UTC',
+    };
+    process.env.WHYD_GENUINE_SIGNUP_SECRET = env.WHYD_GENUINE_SIGNUP_SECRET; // required by ./api-client.js
+    t.context.serverProcess = await startOpenwhydServer(env);
+  } else {
+    process.env.WHYD_GENUINE_SIGNUP_SECRET = 'whatever'; // required by ./api-client.js
+    await refreshOpenwhydCache();
+  }
+
   t.context.openwhyd = require('./api-client');
   t.context.getUser = (id) =>
     testDataCollections.user.find(({ _id }) => id === _id.toString());
 });
 
 test.after((t) => {
-  t.context.serverProcess.kill();
+  if (t.context.serverProcess) {
+    t.context.serverProcess.kill();
+  }
 });
 
 const personas = [
